[func-refs] Check value types for valid ref types
This is a pretty big diff because we have to pass resources to
push_operand now. Also, we have to add yet another method to the
WasmModuleResources trait. Further, there is some pretty hacky error
message mapping going on.

The design up until now has pretty much
assumed all value types are valid, which is no longer true. This commit
is pretty ugly as a result. A more thorough refactoring is necessary,
which i also tried but ended up extremely stuck on and figured i'd defer
to someone more knowledgable (potentially future me)