[func-refs] Assume matches V128 V128 = true
It looks to me that the function references spec was forked from
webassembly before vectors were merged, so it has no definitive answer
on V128 subtyping.  It also means we accidentally reverted `select`
to an outdated spec.  I think for now it's reasonable to use common
sense for V128 subtyping, and at some point the proposal is presumably
planned to be rebased?